chore: Factor out multiple endpoints into a common runner
This creates an Endpoint interface wrapping both the gRPC and the
REST endpoints, as well as the cmux multiplexer that redirects from
the external port to either of those.

This also sets up the infrastructure for graceful shutdown upon
receipt of a signal, though we are waiting for a PR in a dependency
for that to work fully.