Prepare data for REST handlers and construct their gorilla/mux path matchers (#505)
This processes HTTP annotations to extract the data that will be needed to generate the REST/proto shim Go source files. As part of that, this parses the URL matching patterns in each HTTP annotation binding, checks for uniqueness of each pattern across all services, and generates the matching URL pattern that will, in subsequent PRs, be passed to gorilla/mux (GitHub, godoc) as the HTTP server dispatcher.

Note that this does not yet parse the "body" part of HTTP annotations. That will happen in a subsequent PR.