Capture URL path variables, and match HTTP verbs (#516)
We now capture URL path variables. These are echoed in the server logs, but not yet populated in the request protos nor propagated to the backend services.

In addition, the URL matchers now get triggered when called with the correct HTTP verb (as configured in the HTTP annotation). Note that if a handled pattern is invoked with the wrong verb, no handler is called and no error is issued. We will have to figure out later how to detect and report/error on this.

In addition:

- We add checks to make sure that no recursive variables are defined in HTTP annotations, since this is actually disallowed by https://cloud.google.com/endpoints/docs/grpc-service-config/reference/rpc/google.api#path-template-syntax. I had missed the fine-print before :-(
- We remove the explicit Go formatting calls because that seems to be happening regardless