feat: Ensure body and URI fields are not duplicated in query params
This adds a compliance_suite_negative_test to ensure HTTP errors are
generated should the query params include parameters included
elsewhere.

This means that for generators to pass the test suite in
compliance_suite_test.json, they must be strict in what they emit and
not duplicate fields in query params (otherwise Showcase responds with
an error and the test fails).

This also adds better HTTP-server error handling, writng to the server
logs and the response body, and setting standard HTTP error code. This
fixes a bunch of TODOs related to us previously always responding with
an HTTPOK