chore: migrate to Actions (#590)
This migrates most of the CircleCI config to GitHub Actions. There are a couple of followups that need to happen: #591 #603

This also reduces the work of the `util/cmd/release` tool by moving the cross compilation of release binaries to an Action job (as many of our other Go projects do).