feat: Enforce checking for optional field presence/absence in test suite requests (#696)
Compliance service requests now include a new proto field, `server_verify` (which should be set on all requests in `compliance_suite.json`). If set, Showcase requires the name and contents of the received request to match
those of a request in the test suite with the same `name`. This tests that transmitted fields were sent correctly, which is useful for testing field presence (optional fields).

Note that this uses the `embed` package, which requires a switch to Go 1.16.

The `compliance_suite.json` file is moved from its old location under `schema/` to a location under `server/` because `embed` does not follow symlinks. The original location under `schema/` becomes a symlink to the new location so that we may continue easily releasing the suite with other `schema/` files.