feat!: Have compliance service respond with full request message (#768)
This helps test, eg for RepeatDataBodyInfo, that top-level fields are
correctly encoded as query params. This is needed for testing REST for
GCE, which does *not* need to encode query params corresponding to
fields in nested submessages.

BREAKING CHANGE: This changes the structure (and hence wire format) of
the proto message google.showcase.v1beta1.RepeatResponse.