fix: remove content_sent from StreamingSequences (#1351)
Currently, server streaming APIs do not resume from the point of failure. Before removing the contentSent property, the server would keep track of the last content that was sent before failing. To get closure to how server streams currently react to failure, we removed the contentSent property. This means that when we retry the request, the server will send the entire content, rather than starting from where it left off.