feat: Get the error details to show up in JSON response
Note two things, though:
1. The details are included as protobuf-encoded versions of the Any
fields in the curl responses. You can decode the payload using a tool
like https://protobuf-decoder.netlify.app/ to verify this. However, on
the server log, the content is shown decoded.
2. We also get additional info that's not part of the standard error
API, because we're returning the superset
https://pkg.go.dev/google.golang.org/api/googleapi#Error. Need to
confirm and maybe broaden the requirement to accept larger fields.